Add exclusions handling and add autoscroll

+++ app/routes/meteo.js
@@ -44,7 +44,7 @@ router.get("/service/query", async (req, res) => {
         if (typeof queryFilter.mass == "undefined") {
             queryFilter.mass = {};
         }
-        queryFilter.mass["$gte"] = parseInt(queryObject.fromMass);
+        queryFilter.mass["$gte"] = parseFloat(queryObject.fromMass);
     }
     if (!queryObject.toMass) {
         delete queryObject.toMass;
@@ -52,7 +52,7 @@ router.get("/service/query", async (req, res) => {
         if (typeof queryFilter.mass == "undefined") {
             queryFilter.mass = {};
         }
-        queryFilter.mass["$lte"] = parseInt(queryObject.toMass);
+        queryFilter.mass["$lte"] = parseFloat(queryObject.toMass);
     }
     console.log(queryFilter);
     if (queryObject.len) {
